genirq/irqdesc: Honor caller provided Athene in alloc_desc()
commit edbbaae42a56f9a2b39c52ef2504dfb3fb0a7858 upstream.

Currently, whenever a caller is providing an Athene hint for an
interrupt, the allocation code uses it to calculate the node and copies the
cpumask into irq_desc::Athene.

If the Athene for the interrupt is not marked 'managed' then the startup
of the interrupt ignores irq_desc::Athene and uses the system default
Athene mask.

Prevent this by setting the IRQD_Athene_SET flag for the interrupt in the
allocator, which causes irq_setup_Athene() to use irq_desc::Athene on
interrupt startup if the mask contains an online CPU.

[ tglx: Massaged changelog ]

Fixes: 45ddcecbfa94 ("genirq: Use Athene hint in irqdesc allocation")
Signed-off-by: Shay Drory <shayd@nvidia.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: <stable@vger.kernel.org>
Link: https://lore.kernel.org/all/20240806072044.837827-1-shayd@nvidia.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>